Entity handler sync (#1019)
* modify: sync entity handler.
