+++ src/CAServer.Application.Contracts/Commons/ShiftChainHelper.cs
@@ -163,12 +163,17 @@ public static class ShiftChainHelper
 
     public static AddressFormat GetAddressFormat(string fromChain, string address)
     {
-        if (address.Split("_").Length == 3 && address.Split("_")[1].Length == 50)
+        if (address.Contains(CommonConstant.Underline) && address.StartsWith(CommonConstant.ELF))
         {
+            if (!IsAelfAddress(AddressHelper.ToShortAddress(address)))
+            {
+                return AddressFormat.NoSupport;
+            }
             if (address.EndsWith(CommonConstant.MainChainId))
             {
                 return AddressFormat.Main;
-            }else if (address.EndsWith(CommonConstant.TDVWChainId) || address.EndsWith(CommonConstant.TDVVChainId))
+            }
+            if (address.EndsWith(CommonConstant.TDVWChainId) || address.EndsWith(CommonConstant.TDVVChainId))
             {
                 return AddressFormat.Dapp;
             }
@@ -178,30 +183,25 @@ public static class ShiftChainHelper
 
         if (IsAelfAddress(address))
         {
-            if (fromChain == CommonConstant.MainChainId)
-            {
-                return AddressFormat.Main;
-            }
-
-            return AddressFormat.Dapp;
+            return fromChain == CommonConstant.MainChainId ? AddressFormat.Main : AddressFormat.Dapp;
         }
 
-        if (address.Length == 42)
+        if (IsEthAddress(address))
         {
             return AddressFormat.ETH;
         }
 
-        if (address.Length == 34)
+        if (IsTrxAddress(address))
         {
             return AddressFormat.TRX;
         }
 
-        if (address.Length == 43 || address.Length == 44)
+        if (IsSolanaAddress(address))
         {
             return AddressFormat.Solana;
         }
 
-        if (address.Length == 48 && (address.StartsWith("EQ") || address.StartsWith("UQ")))
+        if (IsTonAddress(address))
         {
             return AddressFormat.TON;
         }
@@ -227,6 +227,11 @@ public static class ShiftChainHelper
             return false;
         }
     }
+
+    private static bool IsEthAddress(string address) => Regex.IsMatch(address, NetworkPatternMap["ETH"]);
+    private static bool IsTrxAddress(string address) => Regex.IsMatch(address, NetworkPatternMap["TRX"]);
+    private static bool IsSolanaAddress(string address) => Regex.IsMatch(address, NetworkPatternMap["Solana"]);
+    private static bool IsTonAddress(string address) => Regex.IsMatch(address, NetworkPatternMap["TON"]);
     
     public static bool VerifyAddress(string chain, string address)
     {
@@ -242,6 +247,8 @@ public static class ShiftChainHelper
         
         return !NetworkPatternMap.ContainsKey(chain) || Regex.IsMatch(address, NetworkPatternMap[chain]);
     }
+    
+    
 }
 
 public class ChainInfo
