+++ src/CAServer.EntityEventHandler.Core/ReportHandler.cs
@@ -1,49 +1,22 @@
-using System;
 using System.Threading.Tasks;
-using AElf.Indexing.Elasticsearch;
 using CAServer.DataReporting.Etos;
-using CAServer.Entities.Es;
-using Microsoft.Extensions.Logging;
-using Newtonsoft.Json;
+using CAServer.EntityEventHandler.Core.Service;
 using Volo.Abp.DependencyInjection;
 using Volo.Abp.EventBus.Distributed;
-using Volo.Abp.ObjectMapping;
 
 namespace CAServer.EntityEventHandler.Core;
 
 public class ReportHandler : IDistributedEventHandler<AccountReportEto>, ITransientDependency
 {
-    private readonly INESTRepository<AccountReportIndex, string> _repository;
-    private readonly IObjectMapper _objectMapper;
-    private readonly ILogger<ReportHandler> _logger;
+    private readonly IReportService _reportService;
 
-    public ReportHandler(INESTRepository<AccountReportIndex, string> repository, IObjectMapper objectMapper,
-        ILogger<ReportHandler> logger)
+    public ReportHandler(IReportService reportService)
     {
-        _repository = repository;
-        _objectMapper = objectMapper;
-        _logger = logger;
+        _reportService = reportService;
     }
 
     public async Task HandleEventAsync(AccountReportEto eventData)
     {
-        try
-        {
-            var index = _objectMapper.Map<AccountReportEto, AccountReportIndex>(eventData);
-            var reportIndex = await _repository.GetAsync(index.Id);
-            if (reportIndex == null)
-            {
-                index.CreateTime = DateTime.UtcNow;
-            }
-
-            await _repository.AddOrUpdateAsync(index);
-            _logger.LogInformation("[AccountReport] account report info handle success, caHash:{caHash}",
-                eventData.CaHash);
-        }
-        catch (Exception e)
-        {
-            _logger.LogError(e, "[AccountReport] handle report account data error, data:{data}",
-                JsonConvert.SerializeObject(eventData));
-        }
+        _ = _reportService.HandleAccountReportAsync(eventData);
     }
 }
