+++ src/CAServer.EntityEventHandler.Core/UserExtraInfoHandler.cs
@@ -32,10 +32,8 @@ public class UserExtraInfoHandler : IDistributedEventHandler<UserExtraInfoEto>, 
         try
         {
             var userInfo = _objectMapper.Map<UserExtraInfoEto, UserExtraInfoIndex>(eventData);
-            _logger.LogDebug("User extra info add or update: {eventData}", JsonConvert.SerializeObject(userInfo));
             await _userExtraInfoRepository.AddOrUpdateAsync(userInfo);
-
-            _logger.LogDebug($"User extra info add or update success: {JsonConvert.SerializeObject(userInfo)}");
+            _logger.LogDebug("User extra info add or update success: {data}", JsonConvert.SerializeObject(userInfo));
         }
         catch (Exception ex)
         {
